+++ assets/__init__.py
@@ -1,2 +1,3 @@
-from assets.ant_nest import AntNest, Ant
+from assets.ant_nest import AntNest
+from assets.ant import Ant
 from assets.food import Food
